requirements.txt
# Core dependencies
langchain>=0.1.0
langgraph>=0.0.40
langchain-openai>=0.0.8
langchain_google_genai>=2.1.9
langchain-community>=0.0.20

# GitHub integration
PyGithub>=1.59.0
requests>=2.31.0

# Slack bot
slack-bolt>=1.18.0
slack-sdk>=3.21.0

# Database
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.0
alembic>=1.12.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0